% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lts_dplyr_adf_beta.R
\name{adf_test_by_group}
\alias{adf_test_by_group}
\title{adf_test_by_group}
\usage{
adf_test_by_group(data, time_var, id_var, measurement_vars)
}
\arguments{
\item{data}{A dataframe containing the time series data.}

\item{time_var}{The name of the time variable.}

\item{id_var}{The name of the unique ID column.}

\item{measurement_vars}{A vector of names of the measurement variables.}
}
\value{
A dataframe with ADF test results for each unique ID and measurement variable, with warnings if NA values were removed.
}
\description{
Perform stationarity tests on grouped time series data using ADF test.
}
