% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportInernalData.R
\docType{data}
\name{catchmentsAndRivers}
\alias{catchmentsAndRivers}
\title{Exporting data which is also used internally in the package}
\format{
An object of class \code{data.frame} with 24752 rows and 19 columns.
}
\usage{
catchmentsAndRivers
}
\description{
Exporting data which is also used internally in the package
}
\keyword{datasets}
