% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{defaultTimeSeries}
\alias{defaultTimeSeries}
\title{This is data to be included in my package}
\description{
This is data to be included in my package
}
\references{
\url{github.com\somaSystems\lifeTimes}
}
\author{
My Name \email{dent.lucas@gmail.com}
}
\keyword{data}
