% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_tsMetaData.R
\name{define_tsMetaData}
\alias{define_tsMetaData}
\title{define_tsMetaData}
\usage{
define_tsMetaData(timeSeriesData = clustR, columsForMetaData = metaData)
}
\arguments{
\item{timeSeriesData}{a dataframe of time series data}

\item{columsForMetaData}{a column name or vector of column names to identify metadata columns}

\item{clustR}{A time series dataset (eg.dataframe)
with variables labelling observations by;
observation ID number (currently = cellNumber),
fieldNumber, Treatment, Row (optional), Column (optional),
and Plate (optional).}
}
\value{
returns a subset of the input data that will be used for metaData of
cross correlations functions
}
\description{
define_tsMetaData
}
\examples{
define_tsMetaData(timeSeriesData = defaultTimeSeries ,
columsForMetaData = c("cellNumber","fieldNumber","Treatment","Row","Column","Plate"))


}
