% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_lts_FunctionCalls.R
\name{dev_lts_FunctionCalls}
\alias{dev_lts_FunctionCalls}
\title{dev_lts_FunctionCalls}
\usage{
dev_lts_FunctionCalls(dev_lts_inputVars)
}
\arguments{
\item{dev_lts_inputVars}{user description of variables taken from "lts_input()", otherwise taken from built in default set of variables.}
}
\value{
dev version of output from all functions
}
\description{
dev_lts_FunctionCalls
}
