% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_lts_internalFunctions_generalised.R
\name{dev_lts_tsToWide}
\alias{dev_lts_tsToWide}
\title{dev_lts_internalFunctions}
\usage{
dev_lts_tsToWide(.lts_variables = NULL)
}
\arguments{
\item{.lts_variables}{user description of variables taken from "lts_input()", otherwise taken from built in default set of variables.}
}
\value{
a list that includes time series data, and strings from user input that map variables in the time series data to input in lifeTimes functions. Eg. which column of dataframe is the unit of "time", which is the categorical variables, and which are the variables to compare when generating CCFs.
}
\description{
dev_lts_internalFunctions
}
