% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lts_diff_by_ID_beta.R
\name{difference_by_unique_id}
\alias{difference_by_unique_id}
\title{difference_by_unique_id}
\usage{
difference_by_unique_id(
  data = NULL,
  id_var = "key_num",
  time_var = "dayOfseason",
  measurement_vars = c("flow_m3s", "rainfall_cm"),
  diff_order = 1
)
}
\arguments{
\item{data}{A dataframe containing the time series data. Defaults to the built-in dataset if not provided.}

\item{id_var}{The name of the unique ID column. Defaults to "key_num".}

\item{time_var}{The name of the time variable. Defaults to "dayOfseason".}

\item{measurement_vars}{A vector of names of the measurement variables. Defaults to `c("flow_m3s", "rainfall_cm")`.}

\item{diff_order}{An integer specifying the order of differencing (e.g., 1 for first difference, 2 for second difference, etc.). Defaults to 1.}
}
\value{
A dataframe with the same layout as the input but with differenced values.
}
\description{
Apply differencing to time series grouped by unique ID.
}
