% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifeTimes-package.R
\docType{package}
\name{lifeTimes-package}
\alias{lifeTimes}
\alias{lifeTimes-package}
\title{lifeTimes: lifeTimes}
\description{
This package uses cross-correlation analysis to study coupling of biological process in time series data.
}
\keyword{internal}
