% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lts_lifeTimesChain.R
\name{lifeTimesChain}
\alias{lifeTimesChain}
\title{lts_lifeTimesChain}
\usage{
lifeTimesChain()
}
\value{
When run without entering an argument,
returns a named list including default timeseries data ($clustR),
and default arguments for tsToWide()
}
\description{
A function that generates default input and calls each of the lifeTimes functions
output is ccf from original time series and calculated differences
}
\examples{
outPutCCF <- lifeTimesChain()



}
