% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportInternalData_lts_catchmentsAndRivers.R,
%   R/lts_catchmentsAndRivers.R
\docType{data}
\name{lts_catchmentsAndRivers}
\alias{lts_catchmentsAndRivers}
\title{Exporting data which is also used internally in the package}
\format{
An object of class \code{data.frame} with 24752 rows and 19 columns.

A data frame with columns:
\describe{
  \item{dayOfseason}{The day of the season.}
  \item{rainfall_cm}{Rainfall in cm.}
  \item{flow_m3s}{Flow in cubic meters per second.}
  \item{season}{Season of observation (e.g., Winter, Spring).}
  \item{catchmentRegion}{Region of the catchment area.}
}
}
\source{
Data source described in readme
}
\usage{
lts_catchmentsAndRivers

lts_catchmentsAndRivers
}
\description{
A dataset containing time series data for catchments and rivers.
}
\examples{
head(lts_catchmentsAndRivers)
}
\keyword{datasets}
