% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lts_ccf_to_df.R
\name{lts_ccf_df}
\alias{lts_ccf_df}
\title{lts_ccf_to_df}
\usage{
lts_ccf_df(.lts_ccflist = lts_ccf_list, .lts_variables = NULL)
}
\arguments{
\item{.lts_ccflist}{output from lts_wide_ts_to_ccf()}

\item{.lts_variables}{output from lts_input()}
}
\description{
lts_ccf_to_df
}
