% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lts_TESTING_cluster_ccf_summaries.R
\name{lts_cluster_ccf_summs}
\alias{lts_cluster_ccf_summs}
\title{lts_cluster_ccf_summs}
\usage{
lts_cluster_ccf_summs(.lts_ccf_with_summs = NULL, .lts_variables = NULL)
}
\arguments{
\item{.lts_ccf_with_summs}{summary statistics off ccfs}

\item{.lts_variables}{variables entered int lts_calc()}
}
\value{
returns ccfs with categorical variable levels assigned by clustering
}
\description{
lts_cluster_ccf_summs
}
