% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lts_get_summary_ccf_stats.R
\name{lts_clusterCCFs}
\alias{lts_clusterCCFs}
\title{lts_get_summary_ccf_stats}
\usage{
lts_clusterCCFs(
  .lts_ccfWithMetaData = lts_ccfWithMetaData_compareBy,
  .lts_variables = NULL,
  .chosenLAGforClustering = modeMaxCorrLAG
)
}
\arguments{
\item{.lts_ccfWithMetaData}{out put from lts_join_metadata_to_ccf}

\item{.lts_variables}{output from lts_input()}

\item{.chosenLAGforClustering}{a chosen lag for clustering summary statistics}
}
\value{

}
\description{
lts_get_summary_ccf_stats
}
