% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lts_clusterOutput_LAGranges.R
\docType{data}
\name{lts_clusterOutput_LAGranges}
\alias{lts_clusterOutput_LAGranges}
\title{Cross correlations output data of rainfall and riverflow, for two different river catchments}
\format{
A list with dataframe with yearly rainfall_cm and riverflow, for the Ash river at Mardock,
and the Thames river at Kingston in the United Kingdom:
\describe{
  \item{season}{summer, autumn, winter, spring}
  \item{catchmentRegion}{Ash at Mardock, or, Thames at Kingston}
  ...
}
}
\source{
\url{https://github.com/somaSystems//}
}
\usage{
lts_clusterOutput_LAGranges
}
\description{
A list with cross correlations calculated from rainfall_cm and riverflow, for the Ash river at Mardock,
and the Thames river at Kingston in the United Kingdom
}
\keyword{datasets}
