% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_lts_TESTSING_patching_lts_clusterPlot.R,
%   R/lts_TESTSING_patching_lts_clusterPlot.R
\name{lts_clustPlot}
\alias{lts_clustPlot}
\title{lts_clusterPlot}
\usage{
lts_clustPlot(
  .lts_output = lts_cluster,
  plotType = c("compoundPlot", "draw_treatmentDendrogram", "plt_dendr",
    "heatmapLagZero", "rawTraces", "clusteredLines")
)

lts_clustPlot(
  .lts_output = lts_cluster,
  plotType = c("compoundPlot", "draw_treatmentDendrogram", "plt_dendr",
    "heatmapLagZero", "rawTraces", "clusteredLines")
)
}
\arguments{
\item{.lts_output}{output from the lifeTimes main workflow}

\item{plotType}{choice from c("compoundPlot","draw_treatmentDendrogram",
"plt_dendr","heatmapLagZero","rawTraces","clusteredLines")}
}
\value{
a plot of features and treatments clustered by CCF at lag zero

a plot of features and treatments clustered by CCF at lag zero
}
\description{
lts_clusterPlot

lts_clusterPlot
}
