% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lts_couplingPlot.R
\name{lts_couplingPlot}
\alias{lts_couplingPlot}
\title{lts_couplingPlot}
\usage{
lts_couplingPlot(.lts_output = NULL)
}
\arguments{
\item{.lts_output}{results of lts_input() function, includes
cross correlation calculations and user input variables}
}
\description{
lts_couplingPlot
}
