% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lts_defaultVariables.R
\docType{data}
\name{lts_defaultVariables}
\alias{lts_defaultVariables}
\title{Default Variables for lifeTimes}
\format{
A list of default variables for lts_in function
}
\usage{
lts_defaultVariables
}
\description{
A structured list of default variables used in the lifeTimes package.
}
\examples{
lts_defaultVariables$lts_time
}
\keyword{datasets}
