% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportInternalData_lts_extractedCells.R,
%   R/lts_extractedCells.R
\docType{data}
\name{lts_extractedCells}
\alias{lts_extractedCells}
\title{Exporting data which is also used internally in the package}
\format{
An object of class \code{data.frame} with 281 rows and 55 columns.

A data frame with columns.
}
\usage{
lts_extractedCells

lts_extractedCells
}
\description{
A dataset containing cell-level data extracted for analysis.
}
\examples{
head(lts_extractedCells)
}
\keyword{datasets}
