% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lts_in.R
\name{lts_in}
\alias{lts_in}
\title{lts_in}
\usage{
lts_in(
  .in_tsData = NULL,
  .in_time = c("dayOfseason"),
  .in_compare_categorical = c("season", "catchmentRegion"),
  .in_plot_measured_variables = FALSE,
  .in_pairedComparisons = list(pair_1 = list(y = "flow_m3s", x = "rainfall_cm")),
  .in_uniqueID_colname = "key_num",
  .in_lagMax = NULL,
  .in_clusterByPortions = FALSE,
  .in_metaData = NULL,
  return_intermediate = FALSE
)
}
\arguments{
\item{.in_tsData}{tidy time series data}

\item{.in_time}{name of the "time" variable}

\item{.in_compare_categorical}{names of categorical or explanatory variables to compare CCFs by. In future releases of this function, if there is only one explanatory variable, it will be possible to include a key of set of ".pairedComparisons".}

\item{.in_plot_measured_variables}{logical parameter, set to TRUE if using one categorical variables and want different CCFs plotted against a single categorical variable.}

\item{.in_pairedComparisons}{a single pair or list of pairs, of names of variables to generate cross correlations for.}

\item{.in_uniqueID_colname}{name of column with unique identifier}

\item{.in_lagMax}{maximum lag in CCFs}

\item{.in_clusterByPortions}{defaults to cluster by mean correlation at mode maximum correlated lag, otherwise clusters by "portion" that each grouping/facet of data represents as a total of category 1.}

\item{.in_metaData}{name of columns with metaData}

\item{return_intermediate}{Logical, return intermediate outputs. Default: FALSE.}
}
\value{
calculated cross correlations, summary statistics and clustering that can be used for classification or plotting
}
\description{
Description of the function's purpose.
}
