% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lts_input.R
\name{lts_input}
\alias{lts_input}
\title{lts_input}
\usage{
lts_input(
  .tsData = NULL,
  .time = NULL,
  .compare_categorical = NULL,
  .plot_measured_variables = FALSE,
  .pairedComparisons = NULL,
  .uniqueID_colname = NULL,
  .lagMax = NULL,
  .clusterByPortions = FALSE,
  .metaData = NULL
)
}
\arguments{
\item{.tsData}{tidy time series data}

\item{.time}{name of the "time" variable}

\item{.compare_categorical}{names of categorical or explanatory variables to compare CCFs by. In future releases of this function, if there is only one explanatory variable, it will be possible to include a key of set of ".pairedComparisons".}

\item{.plot_measured_variables}{logical parameter, set to TRUE if using one categorical variables and want different CCFs plotted against a single categorical variable.}

\item{.pairedComparisons}{a single pair or list of pairs, of names of variables to generate cross correlations for.}

\item{.uniqueID_colname}{name of column with unique identifier}

\item{.lagMax}{maximum lag in CCFs}

\item{.clusterByPortions}{defaults to cluster by mean correlation at mode maximum correlated lag, otherwise clusters by "portion" that each grouping/facet of data represents as a total of category 1.}

\item{.metaData}{name of columns with metaData}
}
\value{
a list that includes time series data, and strings from user input that map variables in the time series data to input in lifeTimes functions. Eg. which column of dataframe is the unit of "time", which is the categorical variables, and which are the variables to compare when generating CCFs.
}
\description{
lts_input
}
