% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lts_leadLagClusterPlot.R
\name{lts_leadLagClusterPlot}
\alias{lts_leadLagClusterPlot}
\title{lts_leadLagClusterPlot}
\usage{
lts_leadLagClusterPlot(
  .lts_output = NULL,
  .removeInstanceOfCategoricalByName = NULL,
  .categoryToRemoveInstanceFrom = NULL
)
}
\arguments{
\item{.lts_output}{results of lts_input() function, includes
cross correlation calculations and user input variables}

\item{.removeInstanceOfCategoricalByName}{name a feature that should be removed from
clustering (eg. because of perfect correlation, high noise, or artefact)}

\item{.categoryToRemoveInstanceFrom}{factor type, to remove factor level from}
}
\description{
lts_leadLagClusterPlot
}
