% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lts_lifeTimesInput.R
\name{lts_lifeTimesInput}
\alias{lts_lifeTimesInput}
\title{lts_lifeTimesInput}
\usage{
lts_lifeTimesInput(
  .tsData = NULL,
  .time = c("dayOfseason"),
  .compare_categorical = c("season", "catchmentRegion"),
  .pairedComparisons = list(pair_1 = list(x = "rainfall_cm", y = "flow_m3s")),
  .uniqueID_colname = "key_num",
  .metaData = NULL
)
}
\arguments{
\item{.tsData}{tidy time series data}

\item{.time}{name of the "time" variable}

\item{.compare_categorical}{names of categorical or explanatory variables to compare CCFs by. In future releases of this function, if there is only one explanatory variable, it will be possible to include a key of set of ".pairedComparisons".}

\item{.pairedComparisons}{a single pair or list of pairs, of names of variables to generate cross correlations for.}

\item{.uniqueID_colname}{name of colum with unique identifier}

\item{.metaData}{name of columns with metaData}
}
\value{
a list that includes time series data, and strings from user input that map variables in the time series data to input in lifeTimes functions. Eg. which column of dataframe is the unit of "time", which is the categorical variables, and which are the variables to compare when generating CCFs.
}
\description{
lts_lifeTimesInput
}
