% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lts_join_metadata_to_ccf.R
\name{lts_metaData_ccf_join}
\alias{lts_metaData_ccf_join}
\title{lts_join_metadata_to_ccf}
\usage{
lts_metaData_ccf_join(.lts_dfccf = lts_dfccf, .lts_variables = NULL)
}
\arguments{
\item{.lts_dfccf}{output from lts_ccf_to_df}

\item{.lts_variables}{user description of variables taken from "lts_input()", otherwise taken from built in default set of variables.}
}
\value{
a list that includes time series data, and strings from user input that map variables in the time series data to input in lifeTimes functions. Eg. which column of dataframe is the unit of "time", which is the categorical variables, and which are the variables to compare when generating CCFs.
}
\description{
lts_join_metadata_to_ccf
}
