% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lts_pairsMaker.R
\name{lts_pairsMaker}
\alias{lts_pairsMaker}
\title{lts_pairs
This function creates input to the ".lts_comparisons" argument of "lts_input". It takes a list of all the columns you are interested in and generates all possible non redundant, non self pairs.}
\usage{
lts_pairsMaker(cols = NULL)
}
\arguments{
\item{cols}{a vector of column names to be compared}
}
\value{
returns a list of elements, where each element is a list of variables that will have CCFs calculated
}
\description{
lts_pairs
This function creates input to the ".lts_comparisons" argument of "lts_input". It takes a list of all the columns you are interested in and generates all possible non redundant, non self pairs.
}
\examples{
 lts_pairsMaker(c("pigeon","ferret","crab"))

}
