% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lts_plot_ClustSum.R
\name{lts_plot_ClustSum}
\alias{lts_plot_ClustSum}
\title{lts_leadLagClusterPlot}
\usage{
lts_plot_ClustSum(
  .lts_output = lts_cluster,
  .lts_chosen_clusterFeature = "posNegDiffmedian_corr_by_lag_range",
  .removeInstanceOfCategoricalByName = NULL,
  .categoryToRemoveInstanceFrom = NULL
)
}
\arguments{
\item{.lts_output}{results of lts_input() function, includes
cross correlation calculations and user input variables}

\item{.lts_chosen_clusterFeature}{a feature from CCF summary statistics (lts_ccf_summaries) that will be used for heatmap clustering}

\item{.removeInstanceOfCategoricalByName}{name a feature that should be removed from
clustering (eg. because of perfect correlation, high noise, or artefact)}

\item{.categoryToRemoveInstanceFrom}{factor type, to remove factor level from}
}
\description{
Function to cluster contitions (eg. treatments) and variables (eg. feature measurements) by difference
in past vs future lags
}
