% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lts_prcomp.R
\name{lts_prcomp}
\alias{lts_prcomp}
\title{lts_prcomp}
\usage{
lts_prcomp(.lts_summary_data = NULL)
}
\arguments{
\item{.lts_summary_data}{output Summary data from the lifeTimes main workflow


"plt_dendr","heatmapLagZero","rawTraces","clusteredLines")}
}
\value{
a plot of features and treatments clustered by CCF at lag zero
}
\description{
lts_prcomp
}
