% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lts_TESTING_leadLagClusterPlot.R
\name{lts_sumClustPlot}
\alias{lts_sumClustPlot}
\title{lts_leadLagClusterPlot}
\usage{
lts_sumClustPlot(
  .lts_output = lts_cluster,
  .lts_chosen_clusterFeature = "posNegDiffmedian_corr_by_lag_range",
  .removeInstanceOfCategoricalByName = NULL,
  .categoryToRemoveInstanceFrom = NULL
)
}
\arguments{
\item{.lts_output}{results of lts_input() function, includes
cross correlation calculations and user input variables}

\item{.removeInstanceOfCategoricalByName}{name a feature that should be removed from
clustering (eg. because of perfect correlation, high noise, or artefact)}

\item{.categoryToRemoveInstanceFrom}{factor type, to remove factor level from}
}
\description{
Function to cluster contitions (eg. treatments) and variables (eg. feature measurements) by difference
in past vs future lags
}
