% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lts_summarise_ccf.R
\name{lts_summarise_ccf}
\alias{lts_summarise_ccf}
\title{lts_summarise_ccf}
\usage{
lts_summarise_ccf(
  .lts_ccfWithMetaData = NULL,
  .lts_variables = NULL,
  .lts_portion = 0.25
)
}
\arguments{
\item{.lts_ccfWithMetaData}{ccf with meta data}

\item{.lts_variables}{variables from user input into lts_calc() or lts_input()}

\item{.lts_portion}{selected portion of the maximum range of lags,
to be used for "portioned' summaries}
}
\value{
returns a list of ccf summaries
}
\description{
lts_summarise_ccf
}
