% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lts_summs_clean.R
\name{lts_summs_clean}
\alias{lts_summs_clean}
\title{lts_summs_clean
function to take summary stats from lts and widen to make "tidy" so that there is a single row per observation
function also generates summary stats for original data
outputs original_summs, ccf_summs, and joined_summs which are a join of original and ccf summaries}
\usage{
lts_summs_clean(.lts_output = NULL)
}
\arguments{
\item{.lts_output}{output from the lifeTimes main workflow

"plt_dendr","heatmapLagZero","rawTraces","clusteredLines")}
}
\value{
original_summs, ccf_summs, and joined_summs which are a join of original and ccf summaries
}
\description{
lts_summs_clean
function to take summary stats from lts and widen to make "tidy" so that there is a single row per observation
function also generates summary stats for original data
outputs original_summs, ccf_summs, and joined_summs which are a join of original and ccf summaries
}
