% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lts_ts_to_wide.R
\name{lts_tsToWide}
\alias{lts_tsToWide}
\title{lts_tsToWide
Converts a tidy data frame into a wide format for time series analysis.
Each time series becomes a vector, uniquely identified by categorical and time variables.}
\usage{
lts_tsToWide(.lts_variables = NULL)
}
\arguments{
\item{.lts_variables}{A list of variables:
- lts_data: A tidy dataframe containing the time series data.
- lts_time: Name of the column representing the time variable.
- lts_uniqueID_colname: Column name containing unique identifiers.
- lts_compare_by: Vector of categorical variable names for grouping.
- lts_pariedComparisons: Variables to pivot and create wide-format data.}
}
\value{
A dataframe in wide format with each time series as a vector.
}
\description{
lts_tsToWide
Converts a tidy data frame into a wide format for time series analysis.
Each time series becomes a vector, uniquely identified by categorical and time variables.
}
\keyword{internal}
