% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lts_wide_ts_to_ccf.R
\name{lts_wide_ts_to_ccf}
\alias{lts_wide_ts_to_ccf}
\title{lts_wide_ts_to_ccf}
\usage{
lts_wide_ts_to_ccf(.lts_cast_ts = NULL, .lts_variables = NULL)
}
\arguments{
\item{.lts_cast_ts}{Output from lts_ts_to_wide()}

\item{.lts_variables}{Output from lts_input()}
}
\value{
Returns a list of cross-correlation results, matching the original structure
}
\description{
lts_wide_ts_to_ccf
}
