% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchmentsAndRivers.R
\docType{data}
\name{catchmentsAndRivers}
\alias{catchmentsAndRivers}
\title{Rainfall and riverflow data for two separate catchment regions.}
\format{
a dataframe with yearly rainfall_cm and riverflow, for the Ash river at Mardock,
and the Thames river at Kingston in the United Kingdom:
\describe{
  \item{season}{summer, autumn, winter, spring}
  \item{catchmentRegion}{Ash at Mardock, or, Thames at Kingston}
  ...
}
}
\source{
\url{https://nrfa.ceh.ac.uk//}
}
\usage{
catchmentsAndRivers
}
\description{
A dataframe with yearly rainfall_cm and riverflow, for the Ash river at Mardock,
and the Thames river at Kingston in the United Kingdom
}
\keyword{datasets}
