% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifeTimesChain.R
\name{lifeTimesChain}
\alias{lifeTimesChain}
\title{lifeTimesChain}
\usage{
lifeTimesChain(timeSeriesData = NULL)
}
\arguments{
\item{timeSeriesData}{a time series dataset with observations, timepoints,
measured features, and labels for features to be coupled.}
}
\value{
When run without entering an argument,
returns a named list including default timeseries data ($clustR),
and default arguments for tsToWide()
}
\description{
A function that generates default input and calls each of the lifeTimes functions
output is ccf from original time series and calculated differences
}
\examples{
outPutCCF <- lifeTimesChain()



}
