% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportInternalData_lts_catchmentsAndRivers.R
\docType{data}
\name{lts_catchmentsAndRivers}
\alias{lts_catchmentsAndRivers}
\title{Exporting data which is also used internally in the package}
\format{
An object of class \code{data.frame} with 24752 rows and 19 columns.
}
\usage{
lts_catchmentsAndRivers
}
\description{
Exporting data which is also used internally in the package
}
\keyword{datasets}
