% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patching_lts_clusterPlot.R
\name{lts_clusterPlot}
\alias{lts_clusterPlot}
\title{lts_clusterPlot}
\usage{
lts_clusterPlot(
  .lts_clusterOutput_LAGranges = lts_clusterOutput_LAGranges,
  .lts_variables = NULL,
  plotType = c("compoundPlot", "draw_treatmentDendrogram", "plt_dendr",
    "heatmapLagZero", "rawTraces", "clusteredLines")
)
}
\arguments{
\item{.lts_clusterOutput_LAGranges}{output from the lifeTimes main workflow}

\item{.lts_variables}{user input mapping data to lifeTimes function arguments}

\item{plotType}{choice from c("compoundPlot","draw_treatmentDendrogram",
"plt_dendr","heatmapLagZero","rawTraces","clusteredLines")}
}
\value{
a plot of features and treatments clustered by CCF at lag zero
}
\description{
lts_clusterPlot
}
