% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lts_couplingPlot.R
\name{lts_couplingPlot}
\alias{lts_couplingPlot}
\title{lts_couplingPlot}
\usage{
lts_couplingPlot(
  .lts_output = NULL,
  .lts_facet_by = category1,
  .lts_colour_by = category
)
}
\arguments{
\item{.lts_output}{results of lts_input() function, includes
cross correlation calculations and user input variables}

\item{.lts_facet_by}{choose from categorical variables, category1, or category2, arguments are withtout "".}

\item{.lts_colour_by}{choose from categorical variables, category1, or category2, arguments are without "".}
}
\description{
lts_couplingPlot
}
