% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lts_couplingPlot.R
\name{lts_couplingPlot}
\alias{lts_couplingPlot}
\title{lts_couplingPlot}
\usage{
lts_couplingPlot(
  .lts_clusterOutput_LAGranges = lts_clusterOutput_LAGranges,
  .lts_variables = lts_variables
)
}
\arguments{
\item{.lts_clusterOutput_LAGranges}{list with dataframe of
CCFs returned from lifeTimesChain()}

\item{.lts_variables}{user defined variables mapping data to function arguments}
}
\description{
lts_couplingPlot
}
