% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportInternalData_lts_extractedCells.R
\docType{data}
\name{lts_extractedCells}
\alias{lts_extractedCells}
\title{Exporting data which is also used internally in the package}
\format{
An object of class \code{data.frame} with 281 rows and 55 columns.
}
\usage{
lts_extractedCells
}
\description{
Exporting data which is also used internally in the package
}
\keyword{datasets}
