% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lts_plot_ccfs.R
\name{lts_plot_ccfs}
\alias{lts_plot_ccfs}
\title{lts_clusterPlot}
\usage{
lts_plot_ccfs(
  .lts_output = lts_cluster,
  plotType = c("compoundPlot", "draw_treatmentDendrogram", "plt_dendr",
    "heatmapLagZero", "rawTraces", "clusteredLines")
)
}
\arguments{
\item{.lts_output}{output from the lifeTimes main workflow}

\item{plotType}{choice from c("compoundPlot","draw_treatmentDendrogram",
"plt_dendr","heatmapLagZero","rawTraces","clusteredLines")}
}
\value{
a plot of features and treatments clustered by CCF at lag zero
}
\description{
lts_clusterPlot
}
