% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wide_ts_to_ccf.R
\name{wide_ts_to_ccf}
\alias{wide_ts_to_ccf}
\title{pairedCCFs}
\usage{
wide_ts_to_ccf(wideIDfeatures)
}
\arguments{
\item{wideIDfeatures}{widened features from original time series dataframe}
}
\value{
df_setOfCCF_forEachFeatureAndCellID
}
\description{
pairedCCFs
}
